stage 6.6: add fpgm pruner
