+++ nni/contrib/compression/pruning/basic_pruner.py
@@ -3,18 +3,16 @@
 
 from __future__ import annotations
 
-from collections import defaultdict
-import functools
 import logging
-from typing import Callable, Dict, List, Literal, overload
+from typing import Dict, List, Literal, overload
 
 import torch
 
-from .tools import _DATA, _METRICS, _MASKS, active_sparse_targets_filter, norm_metrics, generate_sparsity, is_active_target
+from .tools import _DATA, _METRICS, _MASKS, active_sparse_targets_filter, norm_metrics, fpgm_metrics, generate_sparsity
 from ..base.compressor import Compressor, Pruner
-from ..base.target_space import TargetType, PruningTargetSpace
+from ..base.target_space import PruningTargetSpace
 from ..base.wrapper import ModuleWrapper
-from ..utils.evaluator import Evaluator, TensorHook
+from ..utils.evaluator import Evaluator
 
 _logger = logging.getLogger(__name__)
 
@@ -70,7 +68,8 @@ class _NormPruner(Pruner):
                     self._remaining_times -= 1
                 debug_msg = f'{self.__class__.__name__} generate masks, remaining times {self._remaining_times}'
                 _logger.debug(debug_msg)
-            if self._current_step == self.interval_steps and (self._remaining_times == 'unlimited' or self._remaining_times > 0):  # type: ignore
+            if self._current_step == self.interval_steps and \
+                (self._remaining_times == 'unlimited' or self._remaining_times > 0):  # type: ignore
                 self._current_step = 0
             if self._is_first_step:
                 self._is_first_step = False
@@ -146,14 +145,13 @@ class L2NormPruner(_NormPruner):
     p = 2
 
 
-class TaylorFOWeightPruner(Pruner):
-    """
-    Taylor FO weight pruner is a pruner which prunes on the first weight dimension by default,
-    based on estimated importance calculated from the first order taylor expansion on weights to achieve a preset level of network sparsity.
-    The estimated importance is defined as the paper
-    `Importance Estimation for Neural Network Pruning <http://jankautz.com/publications/Importance4NNPruning_CVPR19.pdf>`__.
+class FPGMPruner(_NormPruner):
+    r"""
+    FPGM pruner prunes the blocks of the weight with the smallest geometric median.
+    FPGM chooses the weight blocks with the most replaceable contribution.
 
-    :math:`\widehat{\mathcal{I}}_{\mathcal{S}}^{(1)}(\mathbf{W}) \triangleq \sum_{s \in \mathcal{S}} \mathcal{I}_{s}^{(1)}(\mathbf{W})=\sum_{s \in \mathcal{S}}\left(g_{s} w_{s}\right)^{2}`
+    For more details, please refer to
+    `Filter Pruning via Geometric Median for Deep Convolutional Neural Networks Acceleration <https://arxiv.org/abs/1811.00250>`__.
 
     Parameters
     ----------
@@ -161,114 +159,12 @@ class TaylorFOWeightPruner(Pruner):
         Model to be pruned.
     config_list
         A list of dict, each dict configure which module need to be pruned, and how to prune.
-    evaluator
-        TODO: {evaluator_docstring}
-    training_steps
-        The step number used to collect gradients, the masks will be generated after training_steps training.
 
     Examples
     --------
         TODO
     """
+    p = 2
 
-    @overload
-    def __init__(self, model: torch.nn.Module, config_list: List[Dict], evaluator: Evaluator, training_steps: int):
-        ...
-
-    @overload
-    def __init__(self, model: torch.nn.Module, config_list: List[Dict], evaluator: Evaluator, training_steps: int,
-                 existed_wrappers: Dict[str, ModuleWrapper]):
-        ...
-
-    def __init__(self, model: torch.nn.Module, config_list: List[Dict], evaluator: Evaluator, training_steps: int,
-                 existed_wrappers: Dict[str, ModuleWrapper] | None = None):
-        super().__init__(model=model, config_list=config_list, evaluator=evaluator,
-                         existed_wrappers=existed_wrappers)
-        self.evaluator: Evaluator
-        self.training_steps = training_steps
-
-        # trigger masks generation when self._current_step == self.training_steps
-        self._current_step = 0
-        # save all target hooks with format {module_name: {target_name: hook}}
-        self.hooks: Dict[str, Dict[str, TensorHook]] = defaultdict(dict)
-
-        # `interval_steps` and `total_times` are used by `register_trigger`.
-        # `interval_steps` is the optimize step interval for generating masks.
-        # `total_times` is the total generation times of masks.
-        self.interval_steps = training_steps
-        self.total_times: int | Literal['unlimited'] = 1
-
-    @classmethod
-    def from_compressor(cls, compressor: Compressor, new_config_list: List[Dict], training_steps: int, evaluator: Evaluator | None = None):
-        return super().from_compressor(compressor, new_config_list, training_steps=training_steps, evaluator=evaluator)
-
-    def _collect_data(self) -> Dict[str, Dict[str, torch.Tensor]]:
-        data = defaultdict(dict)
-        for module_name, hooks in self.hooks.items():
-            for target_name, hook in hooks.items():
-                if len(hook.buffer) > 0:
-                    data[module_name][target_name] = hook.buffer[0] / self.training_steps
-        return data
-
-    def _calculate_metrics(self, data: Dict[str, Dict[str, torch.Tensor]]) -> _METRICS:
-        return norm_metrics(p=1, data=data, target_spaces=self._target_spaces)
-
-    def _generate_sparsity(self, metrics: _METRICS) -> _MASKS:
-        return generate_sparsity(metrics, self._target_spaces)
-
-    def _register_hooks(self, evaluator: Evaluator):
-        def collector(buffer: List, target: torch.Tensor) -> Callable[[torch.Tensor], None]:
-            # a factory function, return a tensor hook function for target
-            assert len(buffer) == 0, 'Buffer pass to taylor pruner collector is not empty.'
-
-            def collect_taylor(grad: torch.Tensor):
-                if len(buffer) == 0:
-                    buffer.append(torch.zeros_like(grad))
-                if self._current_step < self.training_steps:
-                    buffer[0] += (target.detach() * grad.detach()).pow(2)
-            return collect_taylor
-
-        hook_list = []
-        for module_name, ts in self._target_spaces.items():
-            for target_name, target_space in ts.items():
-                if is_active_target(target_space):
-                    # TODO: add input/output
-                    if target_space.type is TargetType.PARAMETER:
-                        hook = TensorHook(target_space.target, target_name, functools.partial(collector, target=target_space.target))  # type: ignore
-                        hook_list.append(hook)
-                        self.hooks[module_name][target_name] = hook
-                    else:
-                        raise NotImplementedError()
-        evaluator.register_hooks(hook_list)
-
-    def _register_trigger(self, evaluator: Evaluator):
-        assert self.interval_steps >= self.training_steps or self.interval_steps < 0
-        self._remaining_times = self.total_times
-
-        def optimizer_task():
-            self._current_step += 1
-            if self._current_step == self.training_steps:
-                masks = self.generate_masks()
-                self.update_masks(masks)
-                if isinstance(self._remaining_times, int):
-                    self._remaining_times -= 1
-                debug_msg = f'{self.__class__.__name__} generate masks, remaining times {self._remaining_times}'
-                _logger.debug(debug_msg)
-            if self._current_step == self.interval_steps and (self._remaining_times == 'unlimited' or self._remaining_times > 0):  # type: ignore
-                self._current_step = 0
-                for _, hooks in self.hooks.items():
-                    for _, hook in hooks.items():
-                        hook.buffer.clear()
-
-        evaluator.patch_optimizer_step(before_step_tasks=[], after_step_tasks=[optimizer_task])
-
-    def compress(self):
-        self.evaluator.bind_model(self.bound_model, self._get_param_names_map())
-        self.compress_fuse(self.evaluator)
-        self.evaluator.train(self.training_steps)
-        self.evaluator.unbind_model()
-        return self.bound_model, self.get_masks()
-
-    def compress_fuse(self, evaluator: Evaluator):
-        self._register_hooks(evaluator)
-        self._register_trigger(evaluator)
+    def _calculate_metrics(self, data: _DATA) -> _METRICS:
+        return fpgm_metrics(p=self.p, data=data, target_spaces=self._target_spaces)
