+++ nni/contrib/compression/pruning/tools/__init__.py
@@ -2,6 +2,6 @@
 # Licensed under the MIT license.
 
 from .collect_data import _DATA, active_sparse_targets_filter
-from .calculate_metrics import _METRICS, norm_metrics, sum_sigmoid_metric, mean_metric
+from .calculate_metrics import _METRICS, norm_metrics, fpgm_metrics
 from .sparse_gen import _MASKS, generate_sparsity
 from .utils import is_active_target
