+++ nni/contrib/compression/pruning/tools/calculate_metrics.py
@@ -42,51 +42,18 @@ def norm_metrics(p: str | int, data: _DATA, target_spaces: _PRUNING_TARGET_SPACE
     return metrics
 
 
-def sum_sigmoid_metric(data: _DATA, target_spaces: _PRUNING_TARGET_SPACES) -> _METRICS:
-    """
-    Calculate the sum of each block of the value in the given data.
-
-    Parameters
-    ----------
-    data
-        {module_name: {target_name: val}}.
-    target_spaces
-        {module_name: {target_name: pruning_target_space}}. Used to get the related scaler for each value in data.
-    """
+def fpgm_metrics(p: str | int, data: _DATA, target_spaces: _PRUNING_TARGET_SPACES) -> _METRICS:
     def reduce_func(t: torch.Tensor) -> torch.Tensor:
-        return t.sum(dim=-1).sigmoid()
+        reshape_data = t.reshape(-1, t.shape[-1])
+        metric = torch.zeros(reshape_data.shape[0]).type_as(reshape_data)
+        for i in range(reshape_data.shape[0]):
+            metric[i] = (reshape_data - reshape_data[i]).norm(p=p, dim=-1).sum()  # type: ignore
+        return metric.reshape(t.shape[:-1])
 
     metrics = defaultdict(dict)
     for module_name, module_data in data.items():
         for target_name, target_data in module_data.items():
             target_space = target_spaces[module_name][target_name]
-            if target_space._scaler is None:
-                metrics[module_name][target_name] = target_data
-            else:
-                metrics[module_name][target_name] = target_space._scaler.shrink(target_data, reduce_func)
-    return metrics
-
-
-def mean_metric(data: _DATA, target_spaces: _PRUNING_TARGET_SPACES) -> _METRICS:
-    """
-    Calculate the mean of each block of the value in the given data.
-
-    Parameters
-    ----------
-    data
-        {module_name: {target_name: val}}.
-    target_spaces
-        {module_name: {target_name: pruning_target_space}}. Used to get the related scaler for each value in data.
-    """
-    def reduce_func(t: torch.Tensor) -> torch.Tensor:
-        return t.mean(dim=-1)
-
-    metrics = defaultdict(dict)
-    for module_name, module_data in data.items():
-        for target_name, target_data in module_data.items():
-            target_space = target_spaces[module_name][target_name]
-            if target_space._scaler is None:
-                metrics[module_name][target_name] = target_data
-            else:
-                metrics[module_name][target_name] = target_space._scaler.shrink(target_data, reduce_func)
+            assert target_space._scaler is not None, 'FPGM metric do not support finegrained sparse pattern.'
+            metrics[module_name][target_name] = target_space._scaler.shrink(target_data, reduce_func)
     return metrics
